moving balance utils to the general utils folder
